requirements.txt
Flask
openai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
schedule
SpeechRecognition
pyaudio
